Move common create_encoding_dict function to keras side to not import old tf code
